Standardise accordion section headings font size
We currently rely on browser defaults for font size on the accordion heading
sections (although the text font size is actually set in the enclosed button
 and the font size in the section headings impacts only the height of the
sections).

This PR sets the font size on the section headings to match our standard font
 sizing for `<h2>` elements so that we don't rely on browser defaults.
However this means that we the height of the accordion sections is reduced
by ~9px on mobile devices.

| OS / browser        | Before | After |
|---------------------|--------|-------|
| iOS 13 / Safari     | 60px   | 50px  |
| iOS 12 / Safari     | 60px   | 49px  |
| Android 10 / Chrome | 58px   | 50px  |
| Android 9 / Chrome  | 58px   | 50px  |

WCAG 2.1 states that a [minimum target size must be 44 by 44 CSS pixels](https://www.w3.org/WAI/WCAG21/Understanding/target-size.html) which this will meet.